Refactors ADS OPCTL OPERATOR.
