+++ ads/opctl/cli.py
@@ -18,9 +18,7 @@ import ads.opctl.operator.cli
 import ads.opctl.spark.cli
 from ads.common import auth as authutil
 from ads.common.auth import AuthType
-from ads.common.object_storage_details import ObjectStorageDetails
 from ads.opctl.cmds import activate as activate_cmd
-from ads.opctl.cmds import apply as apply_cmd
 from ads.opctl.cmds import cancel as cancel_cmd
 from ads.opctl.cmds import configure as configure_cmd
 from ads.opctl.cmds import deactivate as deactivate_cmd
@@ -38,7 +36,6 @@ from ads.opctl.constants import (
     RESOURCE_TYPE,
     RUNTIME_TYPE,
 )
-from ads.opctl.decorator.common import with_auth
 from ads.opctl.utils import build_image as build_image_cmd
 from ads.opctl.utils import publish_image as publish_image_cmd
 from ads.opctl.utils import suppress_traceback
@@ -726,77 +723,7 @@ def predict(**kwargs):
     suppress_traceback(kwargs["debug"])(predict_cmd)(**kwargs)
 
 
-@commands.command()
-@click.help_option("--help", "-h")
-@click.option("--debug", "-d", help="Set debug mode", is_flag=True, default=False)
-@click.option(
-    "--file", "-f", help="The path to resource YAML file.", required=True, default=None
-)
-@click.option(
-    "--backend",
-    "-b",
-    help=(
-        "Backend name or the path to the operator's backend config YAML file. "
-        f"Example 1: `ads opctl apply -f operator.yaml -b {BACKEND_NAME.LOCAL.value}` "
-        "Supported backends: "
-        f"{[BACKEND_NAME.JOB.value,BACKEND_NAME.DATAFLOW.value,BACKEND_NAME.LOCAL.value,]} "
-        "Example 2: `ads opctl apply -f operator.yaml -b backend.yaml` "
-        "Use the `ads opctl operator init` command to generate operator's configs. "
-    ),
-    required=False,
-    default=None,
-)
-@click.option(
-    "--ads-config",
-    "-c",
-    help="The folder where the ADS opctl config.ini located. Default value: `~/.ads_ops`.",
-    required=False,
-    default=None,
-)
-@click.option(
-    "--dry-run",
-    "-r",
-    default=False,
-    is_flag=True,
-    help="During dry run, the actual operation is not performed, only the steps are enumerated.",
-)
-@click.option(
-    "--auth",
-    "-a",
-    help=(
-        "The authentication method to leverage OCI resources. "
-        "The default value will be taken from the ADS `config.ini` file. "
-        "Check the `ads opctl configure --help` command to get details about the `config.ini`."
-    ),
-    type=click.Choice(AuthType.values()),
-    default=None,
-)
-@with_auth
-def apply(debug: bool, **kwargs: Dict[str, Any]) -> None:
-    """
-    Runs the operator with the given specification on the targeted backend.
-    """
-    operator_spec = {}
-    backend = kwargs.pop("backend")
-
-    auth = {}
-    if any(ObjectStorageDetails.is_oci_path(uri) for uri in (kwargs["file"], backend)):
-        auth = authutil.default_signer()
-
-    with fsspec.open(kwargs["file"], "r", **auth) as f:
-        operator_spec = suppress_traceback(debug)(yaml.safe_load)(f.read())
-
-    if backend and backend.lower().endswith((".yaml", ".yml")):
-        with fsspec.open(backend, "r", **auth) as f:
-            backend = suppress_traceback(debug)(yaml.safe_load)(f.read())
-
-    suppress_traceback(debug)(apply_cmd)(
-        config=operator_spec, backend=backend, **kwargs
-    )
-
-
 commands.add_command(ads.opctl.conda.cli.commands)
 commands.add_command(ads.opctl.model.cli.commands)
 commands.add_command(ads.opctl.spark.cli.commands)
 commands.add_command(ads.opctl.distributed.cli.commands)
-commands.add_command(ads.opctl.operator.cli.commands)
