+++ ads/opctl/operator/cmd.py
@@ -8,7 +8,6 @@ import os
 import re
 import runpy
 import shutil
-import sys
 import tempfile
 from typing import Any, Dict, Union
 
@@ -22,6 +21,9 @@ from ads.common.decorator.runtime_dependency import (
     runtime_dependency,
 )
 from ads.opctl import logger
+from ads.opctl.backend.ads_dataflow import DataFlowOperatorBackend
+from ads.opctl.backend.ads_ml_job import MLJobOperatorBackend
+from ads.opctl.backend.local import LocalOperatorBackend
 from ads.opctl.cmds import _BackendFactory
 from ads.opctl.conda.cmds import create as conda_create
 from ads.opctl.conda.cmds import publish as conda_publish
@@ -320,7 +322,6 @@ def init(
 @validate_environment
 def build_image(
     name: str = None,
-    gpu: bool = None,
     rebuild_base_image: bool = None,
     **kwargs: Dict[str, Any],
 ) -> None:
@@ -331,8 +332,6 @@ def build_image(
     ----------
     name: (str, optional)
         Name of the operator to build the image.
-    gpu: (bool, optional)
-        Whether to build a GPU-enabled Docker image.
     rebuild_base_image: (optional, bool)
         If rebuilding both base and operator's images required.
     kwargs: (Dict, optional).
@@ -354,7 +353,7 @@ def build_image(
     logger.info(f"Building Docker image for the `{operator_info.name}` operator.")
 
     # checks if GPU base image needs to be used.
-    gpu = operator_info.gpu or gpu
+    gpu = operator_info.gpu
 
     cur_dir = os.path.dirname(os.path.abspath(__file__))
     base_image_name = OPERATOR_BASE_GPU_IMAGE if gpu else OPERATOR_BASE_IMAGE
@@ -661,3 +660,123 @@ def create(
         Additional key value arguments.
     """
     raise NotImplementedError()
+
+
+def apply(config: Dict, backend: Union[Dict, str] = None, **kwargs) -> None:
+    """
+    Runs the operator with the given specification on the targeted backend.
+
+    Parameters
+    ----------
+    config: Dict
+        The operator's config.
+    backend: (Union[Dict, str], optional)
+        The backend config or backend name to run the operator.
+    kwargs: (Dict, optional)
+        Optional key value arguments to run the operator.
+    """
+    p = ConfigProcessor(config).step(ConfigMerger, **kwargs)
+
+    if p.config.get("kind", "").lower() != "operator":
+        raise RuntimeError("Not supported kind of workload.")
+
+    from ads.opctl.operator import cmd as operator_cmd
+    from ads.opctl.operator.common.operator_loader import (
+        OperatorInfo,
+        OperatorLoader,
+    )
+
+    operator_type = p.config.get("type", "").lower()
+
+    # validation
+    if not operator_type:
+        raise ValueError(
+            f"The `type` attribute must be specified in the operator's config."
+        )
+
+    # extracting details about the operator
+    operator_info: OperatorInfo = OperatorLoader.from_uri(uri=operator_type).load()
+
+    supported_backends = (
+        BACKEND_NAME.JOB.value,
+        BACKEND_NAME.DATAFLOW.value,
+        BACKEND_NAME.OPERATOR_LOCAL.value,
+        BACKEND_NAME.LOCAL.value,
+    )
+
+    backend_runtime_map = {
+        BACKEND_NAME.JOB.value.lower(): (
+            BACKEND_NAME.JOB.value.lower(),
+            RUNTIME_TYPE.PYTHON.value.lower(),
+        ),
+        BACKEND_NAME.DATAFLOW.value.lower(): (
+            BACKEND_NAME.DATAFLOW.value.lower(),
+            RUNTIME_TYPE.DATAFLOW.value.lower(),
+        ),
+        BACKEND_NAME.OPERATOR_LOCAL.value.lower(): (
+            BACKEND_NAME.OPERATOR_LOCAL.value.lower(),
+            RUNTIME_TYPE.PYTHON.value.lower(),
+        ),
+    }
+
+    if not backend:
+        logger.info(
+            f"Backend config is not provided, the {BACKEND_NAME.LOCAL.value} "
+            "will be used by default. "
+        )
+        backend = {"kind": BACKEND_NAME.OPERATOR_LOCAL.value}
+
+    if isinstance(backend, str):
+        backend = {
+            "kind": BACKEND_NAME.OPERATOR_LOCAL.value
+            if backend.lower() == BACKEND_NAME.LOCAL.value
+            else backend
+        }
+
+    backend_kind = backend.get("kind").lower() or "unknown"
+
+    # If backend kind is Job, then it is necessary to check the infrastructure kind.
+    # This is necessary, because Jobs and DataFlow have similar kind,
+    # The only difference would be in the infrastructure kind.
+    # This is a temporary solution, the logic needs to be placed in the ConfigMerger instead.
+    if backend_kind == BACKEND_NAME.JOB.value:
+        if (
+            backend.get("spec", {}).get("infrastructure", {}).get("type", "").lower()
+            == BACKEND_NAME.DATAFLOW.value
+        ):
+            backend_kind = BACKEND_NAME.DATAFLOW.value
+
+    if backend_kind not in supported_backends:
+        raise RuntimeError(
+            f"Not supported backend - {backend_kind}. Supported backends: {supported_backends}"
+        )
+
+    # generate backend specification in case if it is not provided
+    if not backend.get("spec"):
+        backends = operator_cmd._init_backend_config(
+            operator_info=operator_info, **kwargs
+        )
+        backend = backends[backend_runtime_map[backend_kind]]
+
+    p_backend = ConfigProcessor(
+        {**backend, **{"execution": {"backend": backend_kind}}}
+    ).step(ConfigMerger, **kwargs)
+
+    p.config["runtime"] = backend
+    p.config["infrastructure"] = p_backend.config["infrastructure"]
+    p.config["execution"] = p_backend.config["execution"]
+
+    if p_backend.config["execution"]["backend"].lower() in [
+        BACKEND_NAME.OPERATOR_LOCAL.value,
+        BACKEND_NAME.LOCAL.value,
+    ]:
+        if kwargs.get("dry_run"):
+            logger.info(
+                "The dry run option is not supported for "
+                "the local backend and will be ignored."
+            )
+        LocalOperatorBackend(config=p.config, operator_info=operator_info).run()
+    elif p_backend.config["execution"]["backend"] == BACKEND_NAME.JOB.value:
+        MLJobOperatorBackend(config=p.config, operator_info=operator_info).run()
+    elif p_backend.config["execution"]["backend"] == BACKEND_NAME.DATAFLOW.value:
+        DataFlowOperatorBackend(config=p.config, operator_info=operator_info).run()
